+++ AngularClient/src/dev-all/web-api.component.ts
@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
 
-import { AppSettings } from "../app-settings/app-settings";
+import { environment } from "../environments/environment";
 import { AppHttpClient } from "forcrowd-backbone";
 
 @Component({
@@ -23,51 +23,51 @@ export class WebApiComponent {
   }
 
   badRequest(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/BadRequestResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/BadRequestResult", null).subscribe();
   }
 
   badRequestMessage(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/BadRequestMessageResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/BadRequestMessageResult", null).subscribe();
   }
 
   conflict(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/ConflictResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/ConflictResult", null).subscribe();
   }
 
   exception(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/ExceptionResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/ExceptionResult", null).subscribe();
   }
 
   internalServerError(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/InternalServerErrorResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/InternalServerErrorResult", null).subscribe();
   }
 
   noContentGet(): void {
-    this.httpClient.get(AppSettings.serviceApiUrl + "/ResultTests/NoContentResult").subscribe(this.handleResponse);
+    this.httpClient.get(environment.serviceApiUrl + "/ResultTests/NoContentResult").subscribe(this.handleResponse);
   }
 
   noContentPost(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/NoContentResult", null).subscribe(this.handleResponse);
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/NoContentResult", null).subscribe(this.handleResponse);
   }
 
   okGet(): void {
-    this.httpClient.get(AppSettings.serviceApiUrl + "/ResultTests/OkResult?message=test").subscribe(this.handleResponse);
+    this.httpClient.get(environment.serviceApiUrl + "/ResultTests/OkResult?message=test").subscribe(this.handleResponse);
   }
 
   okPost(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/OkResult", { message: "ok - test" }).subscribe(this.handleResponse);
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/OkResult", { message: "ok - test" }).subscribe(this.handleResponse);
   }
 
   modelStateError(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/ModelStateErrorResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/ModelStateErrorResult", null).subscribe();
   }
 
   notFound(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/NotFoundResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/NotFoundResult", null).subscribe();
   }
 
   unauthorized(): void {
-    this.httpClient.post(AppSettings.serviceApiUrl + "/ResultTests/UnauthorizedResult", null).subscribe();
+    this.httpClient.post(environment.serviceApiUrl + "/ResultTests/UnauthorizedResult", null).subscribe();
   }
 
   private handleResponse(response: Response) {