+++ AngularClient/src/main/core/core.module.ts
@@ -1,6 +1,6 @@
-import { APP_INITIALIZER, NgModule } from "@angular/core";
-import { BrowserModule } from "@angular/platform-browser";
+import { NgModule } from "@angular/core";
 import { FlexLayoutModule } from "@angular/flex-layout";
+import { BrowserModule } from "@angular/platform-browser";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { RouterModule, Routes } from "@angular/router";
 import { Angulartics2Module } from "angulartics2";
@@ -8,7 +8,7 @@ import { Angulartics2Module } from "angulartics2";
 import { ForcrowdBackboneModule, ISettings } from "forcrowd-backbone";
 import { SharedModule } from "../shared/shared.module";
 
-import { environment } from "../../app-settings/environments/environment-settings";
+import { environment } from "../../environments/environment";
 
 // Components
 import { ContributorsComponent } from "./components/contributors.component";
@@ -22,7 +22,6 @@ import { SearchComponent } from "./components/search.component";
 import { AuthGuard } from "./auth-guard.service";
 import { CanDeactivateGuard } from "./can-deactivate-guard.service";
 import { DynamicTitleResolve } from "./dynamic-title-resolve.service";
-import { GoogleAnalyticsService } from "./google-analytics.service";
 import { ProjectService } from "./project.service";
 
 export { AuthGuard, CanDeactivateGuard, DynamicTitleResolve, ProjectService }
@@ -43,12 +42,6 @@ const coreRoutes: Routes = [
 const settings: ISettings = {
   serviceApiUrl: environment.serviceApiUrl,
   serviceODataUrl: environment.serviceODataUrl
-}
-
-export function appInitializer(googleAnalyticsService: GoogleAnalyticsService) {
-  return () => {
-    googleAnalyticsService.configureTrackingCode();
-  };
 }
 
 @NgModule({
@@ -75,17 +68,9 @@ export function appInitializer(googleAnalyticsService: GoogleAnalyticsService) {
     ForcrowdBackboneModule.configure(settings)
   ],
   providers: [
-    // Application initializer
-    {
-      deps: [GoogleAnalyticsService],
-      multi: true,
-      provide: APP_INITIALIZER,
-      useFactory: appInitializer
-    },
     AuthGuard,
     CanDeactivateGuard,
     DynamicTitleResolve,
-    GoogleAnalyticsService,
     ProjectService,
     FlexLayoutModule,
   ]