+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Changelog
 
+**0.9.6**
+* AngularClient
+  * Google Analytics fix
+  * AppSettings was replaced environment
+
 **0.9.5**
 * AngularClient
   * Client + library separation